backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0

# HTTP / Utilities
pydantic==2.9.2
httpx==0.27.2
requests==2.32.3
python-multipart==0.0.12
pyyaml==6.0.2
beautifulsoup4==4.12.3

# AI Providers
openai==1.51.2
google-generativeai==0.8.3
anthropic==0.37.1

# Image / Export
Pillow==11.0.0
openpyxl==3.1.5
jinja2==3.1.4
